Move landing page to separate package
